Update original blog post: remove names from end.
This is becoming an unhealthy pattern where we're removing names as people
leave =D, so nipping this bud in the root.

Another way to look at it: I'm removing both Daniel and Aseem together. ;)

We should have used our personal blogs from the start for mentioning names,
which is thankfully what we're doing now. =) The official blog can now be
simple and focused on the product.